requirements.txt
-i https://pypi.tuna.tsinghua.edu.cn/simple
fastapi
uvicorn
pillow
numpy